+++ History.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+1.0.2 / 2014-05-22
+==================
 
   * Handle `req.body` key referencing array or object
   * Handle multiple HTTP headers